Add offer details rating bar

+++ app/src/main/java/com/gpetuhov/android/hive/domain/model/Offer.kt
@@ -10,7 +10,7 @@ data class Offer(
     var isFree: Boolean,
     var isActive: Boolean,
     var isFavorite: Boolean = false,
-    var rating: Double = 0.0
+    var rating: Float = 0.0F
 ) {
     var photoList = mutableListOf<Photo>()
 }
